v2.0.2 -A
+ Got rid of confusing and similar variable names being called for tag arguments.

+ Added shorthand and support for lists (ol + ul).

+ Added the BeautifulSoup4 HTML formatter, which is now built in to the Transpiler;
and is disabled by default.  The enable, type -f True


x All $commands and tags require the syntax separator to take values.
This will be changed in the next fewer versions.

x The child element is unable to receive tag attributes as of yet.
This is a low priority fix.


? I want to refactor the custom ERROR function, and make it
easier to read and use.