v2.1.0 -A
[Very big update.]

+ Removed separate functions to change Syntax separator functions & have 
replaced them with the them with the '@set' function.

+ Changed the the start character for functions from '$' to '@'.
Variables will now use the '$' symbol at the prefix.

+ Added variables to the language, and the user can declare and use them
in scripts.  I plan on adding environment variables next.
You can declare ints and strings


x All @commands and tags require the syntax separator to take values.
This will be changed in the next couple of versions.

x The child element is unable to receive tag attributes as of yet.
This is a low priority fix.


? Variables can only be used in the text display as of now, this does not include
lists either.  I plan to update this so you can use variables anywhere within tags.

? I want to add more HTML tags.

? I want to clean and refactor code where it is needed.
Including better commenting.

? I want to update the github Readme.md description.

? I want to create a user manual for the language.